assignments/advnetsci-node-degree/requirements.txt
marimo>=0.14.16
python-igraph>=0.11.9
numpy>=2.2.6
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn
networkx
ipykernel
tqdm
black
nbconvert
jupytext
